Fix Read Multi Renew Cache Bug
- Bug:
When the context {renew_cache: true} is passed to a cached query in an
application that is using a cache_store that supports 'read_multi' (e.g.
Redis) the cache is not cleared and {renew_cache: true} is ignored.

- Expected Behavior:
When the context {renew_cache: true} is passed to a cached query the
cache is renewed.

This commit fixes the above issue by adding logic that checks for
renew_cache: true in the context of fragments passed to the read_multi
method and returns nil for their value so that cache is refreshed. With
this change their is feature parity between cache_stores that support
read_multi and those that don't.