Fix warning on ActiveSupport::Cache.format_version
`Rail.application` was always evaluating to `nil` before as the code wasn't
running in an initializer.

As a fix, instead simply provide a non-deprecated serializer to the
initialized cache_store when in test environment (and where the code
would raise a warning otherwise). Also move this block to a proper Rails
initializer.